nara_service/nara_mcp/requirements.txt
fastmcp>=0.2.0
python-dotenv>=1.0.0